Make using macros nice
